requirements.txt
keyring
pycrypto
pymongo